Extract cookie setting logic into separate function (MT-6).
We have two places where we have to set cookie:

- Login with login/password
- Login with Google
