--- backend/config/__init__.py
@@ -1,2 +1,0 @@
-# Import schema so OpenAPI extensions can register themselves
-import config.schema  # noqa
