--- backend/config/schema.py
@@ -1,1 +1,0 @@
-from login.openapi_custom_auth_extension import CustomJWTAuthenticationExtension  # noqa
